+++ src/professions/admin/add-profession/top-level-information.controller.spec.ts
@@ -1,6 +1,7 @@
 import { DeepMocked, createMock } from '@golevelup/ts-jest';
 import { TestingModule, Test } from '@nestjs/testing';
 import { Response } from 'express';
+import { I18nService } from 'nestjs-i18n';
 import { IndustriesService } from '../../../industries/industries.service';
 import { Industry } from '../../../industries/industry.entity';
 import { Profession } from '../../profession.entity';
@@ -13,6 +14,7 @@ describe('TopLevelInformationController', () => {
   let industriesService: DeepMocked<IndustriesService>;
   let response: DeepMocked<Response>;
   let profession: DeepMocked<Profession>;
+  let i18nService: DeepMocked<I18nService>;
 
   const healthIndustry = new Industry('industries.health');
   healthIndustry.id = 'health-uuid';
@@ -35,11 +37,33 @@ describe('TopLevelInformationController', () => {
       find: async () => profession,
     });
 
+    i18nService = createMock<I18nService>();
+
+    i18nService.translate.mockImplementation(async (text) => {
+      switch (text) {
+        case 'industries.health':
+          return 'Health';
+        case 'industries.constructionAndEngineering':
+          return 'Construction & Engineering';
+        case 'nations.england':
+          return 'England';
+        case 'nations.scotland':
+          return 'Scotland';
+        case 'nations.wales':
+          return 'Wales';
+        case 'nations.northernIreland':
+          return 'Northern Ireland';
+        default:
+          return '';
+      }
+    });
+
     const module: TestingModule = await Test.createTestingModule({
       controllers: [TopLevelInformationController],
       providers: [
         { provide: IndustriesService, useValue: industriesService },
         { provide: ProfessionsService, useValue: professionsService },
+        { provide: I18nService, useValue: i18nService },
       ],
     }).compile();
 
@@ -65,34 +89,34 @@ describe('TopLevelInformationController', () => {
           name: null,
           industriesCheckboxArgs: [
             {
-              text: 'industries.health',
+              text: 'Health',
               value: 'health-uuid',
               checked: false,
             },
             {
-              text: 'industries.constructionAndEngineering',
+              text: 'Construction & Engineering',
               value: 'construction-uuid',
               checked: false,
             },
           ],
           nationsCheckboxArgs: [
             {
-              text: 'nations.england',
+              text: 'England',
               value: 'GB-ENG',
               checked: false,
             },
             {
-              text: 'nations.scotland',
+              text: 'Scotland',
               value: 'GB-SCT',
               checked: false,
             },
             {
-              text: 'nations.wales',
+              text: 'Wales',
               value: 'GB-WLS',
               checked: false,
             },
             {
-              text: 'nations.northernIreland',
+              text: 'Northern Ireland',
               value: 'GB-NIR',
               checked: false,
             },
@@ -129,34 +153,34 @@ describe('TopLevelInformationController', () => {
           name: 'Example Profession',
           industriesCheckboxArgs: [
             {
-              text: 'industries.health',
+              text: 'Health',
               value: 'health-uuid',
               checked: true,
             },
             {
-              text: 'industries.constructionAndEngineering',
+              text: 'Construction & Engineering',
               value: 'construction-uuid',
               checked: false,
             },
           ],
           nationsCheckboxArgs: [
             {
-              text: 'nations.england',
+              text: 'England',
               value: 'GB-ENG',
               checked: true,
             },
             {
-              text: 'nations.scotland',
+              text: 'Scotland',
               value: 'GB-SCT',
               checked: true,
             },
             {
-              text: 'nations.wales',
+              text: 'Wales',
               value: 'GB-WLS',
               checked: false,
             },
             {
-              text: 'nations.northernIreland',
+              text: 'Northern Ireland',
               value: 'GB-NIR',
               checked: false,
             },