+++ src/professions/search/interfaces/profession-search-result-template.interface.ts
@@ -1,0 +1,6 @@
+export interface ProfessionSearchResultTemplate {
+  name: string;
+  slug: string;
+  nations: string;
+  industries: string[];
+}