+++ CHANGELOG.md
@@ -6,4 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- A professions search page that allows for filtering professions by keywords, nations, and industries
+
+### Changed
+
 - Create draft Professions in the database, rather than using session storage when adding a new profession